refactor: 더미 데이터 import 수정
